feat(feature-setup): add template search and apply flow
Implement menu template search with keyboard submit trigger,
template apply with full details API fetch, and fix Double
parsing for server decimal price responses in DTOs.

+++ core/core-network/src/main/kotlin/com/team/chord/core/network/dto/menu/MenuDtos.kt
@@ -19,7 +19,7 @@ data class SearchMenusDto(
 data class TemplateBasicDto(
     val templateId: Long,
     val menuName: String,
-    val defaultSellingPrice: Int,
+    val defaultSellingPrice: Double,
     val categoryCode: String,
     val workTime: Int,
 )
@@ -27,8 +27,8 @@ data class TemplateBasicDto(
 @Serializable
 data class RecipeTemplateDto(
     val ingredientName: String,
-    val defaultUsageAmount: Int,
-    val defaultPrice: Int,
+    val defaultUsageAmount: Double,
+    val defaultPrice: Double,
     val unitCode: String,
 )
